Refactor Monitor classes into partial files
