+++ Sources/DesktopManager/MonitorNativeMethods.Shell.cs
@@ -1,0 +1,38 @@
+using System;
+using System.Runtime.InteropServices;
+using System.Text;
+
+namespace DesktopManager;
+
+public static partial class MonitorNativeMethods
+{
+    [DllImport("user32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
+    public static extern bool SystemParametersInfo(uint uiAction, uint uiParam, StringBuilder pvParam, uint fWinIni);
+
+    [DllImport("user32.dll", SetLastError = true, CharSet = CharSet.Unicode)]
+    public static extern bool SystemParametersInfo(uint uiAction, uint uiParam, string pvParam, uint fWinIni);
+
+    public const uint SPI_SETDESKWALLPAPER = 0x0014;
+    public const uint SPI_GETDESKWALLPAPER = 0x0073;
+    public const uint SPIF_UPDATEINIFILE = 0x0001;
+    public const uint SPIF_SENDWININICHANGE = 0x0002;
+
+    [DllImport("shell32.dll", CharSet = CharSet.Unicode, SetLastError = true)]
+    public static extern int SHCreateItemFromParsingName(string pszPath, IntPtr pbc, ref Guid riid, out IntPtr ppv);
+
+    [ComImport, Guid("92CA9DCD-5622-4bba-A805-5E9F541BD8C9"), InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+    public interface IObjectArray
+    {
+        uint GetCount();
+        [return: MarshalAs(UnmanagedType.Interface)] object GetAt(uint uiIndex, ref Guid riid);
+    }
+
+    [ComImport, Guid("5632b1a4-e38a-400a-928a-d4cd63230295"), InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+    public interface IObjectCollection : IObjectArray
+    {
+        void AddObject([MarshalAs(UnmanagedType.Interface)] object pv);
+        void AddFromArray(IObjectArray poaSource);
+        void RemoveObjectAt(uint uiIndex);
+        void Clear();
+    }
+}
