+++ Sources/DesktopManager/MonitorService.Display.cs
@@ -1,7 +1,6 @@
 using System;
 using System.IO;
 using System.Linq;
-using System.Net.Http;
 using System.Runtime.InteropServices;
 using System.Text;
 using System.Collections.Generic;
@@ -9,286 +8,7 @@ using Microsoft.Win32;
 
 namespace DesktopManager;
 
-/// <summary>
-/// Service for managing monitors and their settings.
-/// </summary>
-public class MonitorService {
-    private const int ENUM_CURRENT_SETTINGS = -1;
-    private readonly IDesktopManager _desktopManager;
-
-    private void Execute(Action action, string operation) {
-        try {
-            action();
-        } catch (COMException ex) {
-            throw new DesktopManagerException(operation, ex);
-        }
-    }
-
-    private T Execute<T>(Func<T> func, string operation) {
-        try {
-            return func();
-        } catch (COMException ex) {
-            throw new DesktopManagerException(operation, ex);
-        }
-    }
-
-    /// <summary>
-    /// Initializes a new instance of the <see cref="MonitorService"/> class.
-    /// </summary>
-    /// <param name="desktopManager">The desktop manager interface.</param>
-    public MonitorService(IDesktopManager desktopManager) {
-        _desktopManager = desktopManager;
-
-        try {
-            Execute(() => _desktopManager.Enable(), nameof(IDesktopManager.Enable));
-        } catch (DesktopManagerException) {
-            // COM failures are ignored during initialization to support unsupported scenarios
-        }
-    }
-
-    /// <summary>
-    /// Gets the list of all monitors.
-    /// </summary>
-    /// <returns>A list of <see cref="Monitor"/> objects.</returns>
-    public List<Monitor> GetMonitors() {
-        try {
-            List<Monitor> list = new List<Monitor>();
-
-            var count = Execute(() => _desktopManager.GetMonitorDevicePathCount(), nameof(IDesktopManager.GetMonitorDevicePathCount));
-            for (uint i = 0; i < count; i++) {
-                var monitor = new Monitor(this) {
-                    Index = (int)i,
-                    DeviceId = Execute(() => _desktopManager.GetMonitorDevicePathAt(i), nameof(IDesktopManager.GetMonitorDevicePathAt))
-                };
-                if (monitor.DeviceId != "") {
-                    monitor.WallpaperPosition = Execute(() => _desktopManager.GetPosition(), nameof(IDesktopManager.GetPosition));
-                    monitor.Wallpaper = Execute(() => _desktopManager.GetWallpaper(monitor.DeviceId), nameof(IDesktopManager.GetWallpaper));
-                    monitor.Rect = Execute(() => _desktopManager.GetMonitorBounds(monitor.DeviceId), nameof(IDesktopManager.GetMonitorBounds));
-
-                    // Populate new properties
-                    DISPLAY_DEVICE d = new DISPLAY_DEVICE();
-                    d.cb = Marshal.SizeOf(d);
-                    if (MonitorNativeMethods.EnumDisplayDevices(null, i, ref d, 0)) {
-                        monitor.DeviceName = d.DeviceName;
-                        monitor.DeviceString = d.DeviceString;
-                        monitor.StateFlags = d.StateFlags;
-                        monitor.DeviceKey = d.DeviceKey;
-                    }
-                }
-                list.Add(monitor);
-            }
-
-            return list;
-        } catch (DesktopManagerException) {
-            return EnumerateMonitorsFallback();
-        } catch (COMException) {
-            return EnumerateMonitorsFallback();
-        }
-    }
-
-    private List<Monitor> EnumerateMonitorsFallback() {
-        List<Monitor> monitors = new List<Monitor>();
-        int index = 0;
-        MonitorNativeMethods.MonitorEnumProc proc = (IntPtr hMonitor, IntPtr hdc, ref RECT rect, IntPtr lparam) => {
-            MONITORINFOEX info = new MONITORINFOEX();
-            info.cbSize = Marshal.SizeOf<MONITORINFOEX>();
-            if (MonitorNativeMethods.GetMonitorInfo(hMonitor, ref info)) {
-                DISPLAY_DEVICE device = new DISPLAY_DEVICE();
-                device.cb = Marshal.SizeOf(device);
-                if (MonitorNativeMethods.EnumDisplayDevices(info.szDevice, 0, ref device, 0)) {
-                    var monitor = new Monitor(this) {
-                        Index = index,
-                        DeviceId = info.szDevice,
-                        DeviceName = device.DeviceName,
-                        DeviceString = device.DeviceString,
-                        StateFlags = device.StateFlags,
-                        DeviceKey = device.DeviceKey,
-                        Rect = info.rcMonitor
-                    };
-                    monitors.Add(monitor);
-                }
-            }
-            index++;
-            return true;
-        };
-        MonitorNativeMethods.EnumDisplayMonitors(IntPtr.Zero, IntPtr.Zero, proc, IntPtr.Zero);
-        return monitors;
-    }
-
-    /// <summary>
-    /// Gets the list of connected monitors.
-    /// </summary>
-    /// <returns>A list of connected <see cref="Monitor"/> objects.</returns>
-    public List<Monitor> GetMonitorsConnected() {
-        List<Monitor> list = new List<Monitor>();
-        foreach (var monitor in GetMonitors()) {
-            if (monitor.DeviceId != "") {
-                list.Add(monitor);
-            }
-        }
-        return list;
-    }
-
-    /// <summary>
-    /// Sets the wallpaper for a specific monitor.
-    /// </summary>
-    /// <param name="monitorId">The monitor ID.</param>
-    /// <param name="wallpaperPath">The path to the wallpaper image.</param>
-    public void SetWallpaper(string monitorId, string wallpaperPath) {
-        try {
-            Execute(() => _desktopManager.SetWallpaper(monitorId, wallpaperPath), nameof(IDesktopManager.SetWallpaper));
-        } catch (DesktopManagerException) {
-            SetSystemWallpaper(wallpaperPath);
-        } catch (COMException) {
-            SetSystemWallpaper(wallpaperPath);
-        }
-    }
-
-    /// <summary>
-    /// Sets the wallpaper for a specific monitor from a data stream.
-    /// </summary>
-    /// <param name="monitorId">The monitor ID.</param>
-    /// <param name="imageStream">Stream containing image data.</param>
-    public void SetWallpaper(string monitorId, Stream imageStream) {
-        string temp = WriteStreamToTempFile(imageStream);
-        try {
-            SetWallpaper(monitorId, temp);
-        } finally {
-            DeleteTempFile(temp);
-        }
-    }
-
-    /// <summary>
-    /// Sets the wallpaper for a specific monitor from a URL.
-    /// </summary>
-    /// <param name="monitorId">The monitor ID.</param>
-    /// <param name="url">URL pointing to the image.</param>
-    public void SetWallpaperFromUrl(string monitorId, string url) {
-        using HttpClient client = new();
-        using Stream stream = client.GetStreamAsync(url).GetAwaiter().GetResult();
-        SetWallpaper(monitorId, stream);
-    }
-
-    /// <summary>
-    /// Sets the wallpaper for a monitor by index.
-    /// </summary>
-    /// <param name="index">The index of the monitor.</param>
-    /// <param name="wallpaperPath">The path to the wallpaper image.</param>
-    public void SetWallpaper(int index, string wallpaperPath) {
-        try {
-            var monitorId = Execute(() => _desktopManager.GetMonitorDevicePathAt((uint)index), nameof(IDesktopManager.GetMonitorDevicePathAt));
-            Execute(() => _desktopManager.SetWallpaper(monitorId, wallpaperPath), nameof(IDesktopManager.SetWallpaper));
-        } catch (DesktopManagerException) {
-            SetSystemWallpaper(wallpaperPath);
-        } catch (COMException) {
-            SetSystemWallpaper(wallpaperPath);
-        }
-    }
-
-    /// <summary>
-    /// Sets the wallpaper for a monitor by index from a data stream.
-    /// </summary>
-    /// <param name="index">The index of the monitor.</param>
-    /// <param name="imageStream">Stream containing image data.</param>
-    public void SetWallpaper(int index, Stream imageStream) {
-        var monitorId = Execute(() => _desktopManager.GetMonitorDevicePathAt((uint)index), nameof(IDesktopManager.GetMonitorDevicePathAt));
-        SetWallpaper(monitorId, imageStream);
-    }
-
-    /// <summary>
-    /// Sets the wallpaper for a monitor by index from a URL.
-    /// </summary>
-    /// <param name="index">The index of the monitor.</param>
-    /// <param name="url">URL pointing to the image.</param>
-    public void SetWallpaperFromUrl(int index, string url) {
-        var monitorId = Execute(() => _desktopManager.GetMonitorDevicePathAt((uint)index), nameof(IDesktopManager.GetMonitorDevicePathAt));
-        SetWallpaperFromUrl(monitorId, url);
-    }
-
-    /// <summary>
-    /// Sets the wallpaper for all connected monitors.
-    /// </summary>
-    /// <param name="wallpaperPath">The path to the wallpaper image.</param>
-    public void SetWallpaper(string wallpaperPath) {
-        try {
-            var devicePathCount = GetMonitorsConnected();
-            foreach (var device in devicePathCount) {
-                Execute(() => _desktopManager.SetWallpaper(device.DeviceId, wallpaperPath), nameof(IDesktopManager.SetWallpaper));
-            }
-        } catch (DesktopManagerException) {
-            SetSystemWallpaper(wallpaperPath);
-        } catch (COMException) {
-            SetSystemWallpaper(wallpaperPath);
-        }
-    }
-
-    /// <summary>
-    /// Sets the wallpaper for all monitors using image data stream.
-    /// </summary>
-    /// <param name="imageStream">Stream containing image data.</param>
-    public void SetWallpaper(Stream imageStream) {
-        string temp = WriteStreamToTempFile(imageStream);
-        try {
-            SetWallpaper(temp);
-        } finally {
-            DeleteTempFile(temp);
-        }
-    }
-
-    /// <summary>
-    /// Sets the wallpaper for all monitors using an image from a URL.
-    /// </summary>
-    /// <param name="url">URL pointing to the image.</param>
-    public void SetWallpaperFromUrl(string url) {
-        using HttpClient client = new();
-        using Stream stream = client.GetStreamAsync(url).GetAwaiter().GetResult();
-        SetWallpaper(stream);
-    }
-
-    /// <summary>
-    /// Gets the wallpaper for a specific monitor.
-    /// </summary>
-    /// <param name="monitorId">The monitor ID.</param>
-    /// <returns>The path to the wallpaper image.</returns>
-    public string GetWallpaper(string monitorId) {
-        try {
-            return Execute(() => _desktopManager.GetWallpaper(monitorId), nameof(IDesktopManager.GetWallpaper));
-        } catch (DesktopManagerException) {
-            return GetSystemWallpaper();
-        } catch (COMException) {
-            return GetSystemWallpaper();
-        }
-    }
-
-    /// <summary>
-    /// Gets the wallpaper for a monitor by index.
-    /// </summary>
-    /// <param name="index">The index of the monitor.</param>
-    /// <returns>The path to the wallpaper image.</returns>
-    public string GetWallpaper(int index) {
-        try {
-            var monitorId = Execute(() => _desktopManager.GetMonitorDevicePathAt((uint)index), nameof(IDesktopManager.GetMonitorDevicePathAt));
-            return Execute(() => _desktopManager.GetWallpaper(monitorId), nameof(IDesktopManager.GetWallpaper));
-        } catch (DesktopManagerException) {
-            return GetSystemWallpaper();
-        } catch (COMException) {
-            return GetSystemWallpaper();
-        }
-    }
-
-    /// <summary>
-    /// Gets the device path of a monitor by index.
-    /// </summary>
-    /// <param name="index">The index of the monitor.</param>
-    /// <returns>The device path of the monitor.</returns>
-    public string GetMonitorDevicePathAt(uint index) {
-        return Execute(() => _desktopManager.GetMonitorDevicePathAt(index), nameof(IDesktopManager.GetMonitorDevicePathAt));
-    }
-
-    /// <summary>
-    /// Gets the desktop background color.
-    /// </summary>
-    /// <returns>The background color as RGB value.</returns>
+public partial class MonitorService {
     public uint GetBackgroundColor() {
         try {
             return Execute(() => _desktopManager.GetBackgroundColor(), nameof(IDesktopManager.GetBackgroundColor));
@@ -789,3 +509,4 @@ public class MonitorService {
         return devices;
     }
 }
+
